incorporate pr feedback
