+++ frontend/src/components/dialogs/GDialog.vue
@@ -163,10 +163,10 @@ export default {
     }
   },
   methods: {
-    confirmWithDialog ({ onShowCallback, confirmCallback } = {}) {
+    confirmWithDialog ({ confirmationInterceptor } = {}) {
       this.showDialog()
       this.userInput = ''
-      this.confirmCallback = confirmCallback
+      this.confirmationInterceptor = confirmationInterceptor
 
       // we must delay the "focus" handling because the dialog.open is animated
       // and the 'autofocus' property didn't work in this case.
@@ -174,12 +174,6 @@ export default {
         setDelayedInputFocus(this, 'deleteDialogInput')
       }
 
-      if (onShowCallback) {
-        this.$nextTick(() => {
-          // delay the callback to make sure that all components are loaded
-          onShowCallback()
-        })
-      }
       return new Promise(resolve => {
         this.resolve = resolve
       })
@@ -213,8 +207,8 @@ export default {
     async resolveAction (value) {
       if (isFunction(this.resolve)) {
         if (value) {
-          if (this.confirmCallback) {
-            const confirmed = await this.confirmCallback()
+          if (this.confirmationInterceptor) {
+            const confirmed = await this.confirmationInterceptor()
             if (!confirmed) {
               // cancel resolve action
               return
