+++ frontend/src/sass/main.scss
@@ -43,6 +43,10 @@
   cursor: pointer;
 }
 
+.full-width {
+  width: 100%;
+}
+
 .mr-extra {
   margin-right: 96px;
 }
